Implement targeted search
When prefixing the search term with a special term (currently `wiki`, `poedb`, `ninja`, `trade` and `tool`) and a colon, the search will only show results matching that kind of term.
This is useful when you aren't just generally interested in an item, but already know _where_ you want to look for it.
Examples:
* If you want to check the price history of Maven's Writs, type `ninja:writ`.
* If you want to know which tags the affixes on The Brass Dome have, type `poedb:brass dome`.
* If you want to buy a Headhunter, type `trade:headhunter`.
* If you want to check the layout of today's Uber Lab, type `tool:eternal`.

In case of a targeted search, allow more result items to be rendere. We have a lot more screen space available in this scenario.

Resolves #3