Change default shortcut to Ctrl+P
The current shortcut Ctrl+Shift+P is too long and hard to press, especially when you use it frequently.
Change it to Ctrl+P. This is the canonical shortcut for printing in many applications, but I doubt many people of the target audience of this application still prints stuff, so I'm not concerned about shadowing it.

Note: Ctrl+Shift+P will most likely make a comeback as default shortcut for #10.