Make palette shortcut configurable in config file
There should be a way to change the palette shortcut. Now is not a good time to make it configurable in the UI, so we'll at least offer the user to change it in the config file.

The config file is located at `%APPDATA&\poe-palette\config.json`.
Check the Electron docs for [available key codes and modifiers](https://www.electronjs.org/docs/latest/api/accelerator#available-modifiers).

Partially addresses #20.