Check for newer app version at start
When the app starts, check for a newer app version. If there is a newer release, offer to open the release page in the browser.
When the user chooses to open the release page, exit the app. The user is likely to download the newer release, removing the necessity to close the old app instance manually first.