Periodically check for newer data in background
Data for the search index is recalculated every 24 hours and made available to the clients via Cloudfront. Previously, the client would only get the latest data when starting the app, meaning users with long-running clients would miss out on the daily updates.
Implement a HEAD check for newer data of the currently selected league and load it when found.

Resolves #47