Craft of Exile integration
This PR adds support for Craft of Exile, both fo the data scraper and the electron app.
Searching for base items (e.g. "Fingerless Slilk Gloves") will now offer the addition result `Craft <base item>` and open that base item in Craft of Exile's calculator.

The Craft of Exile integration is disabled by default and must be activated in the integration selector.

Resolves #40