Add header image to help with visual clarity
This PR adds a header image above the search bar to help with visual clarity of the palette.
Previously, opening the palette above dark-ish backgrounds made it hard to easily spot its position.
The addition of the brighter header image (that is directly attached to the search field) helps with that.